Added typing and mypy changes

+++ BallCollisionSimulator.py
@@ -1,65 +1,65 @@
-import math
 import readchar
 import argparse
 from dataclasses import dataclass
 from copy import copy
 import vpython as vp
+from typing import Tuple, Optional
 
 
 class PhysicsParameters:
-    def __init__(self, mass, position, velocity):
-        self.mass = mass
-        self.position = vp.vector(*position, 0)
-        self.velocity = vp.vector(*velocity, 0)
+    def __init__(self, mass: float, position: Tuple[float, float], velocity: Tuple[float, float]):
+        self.mass: float = mass
+        self.position: vp.vector = vp.vector(*position, 0)
+        self.velocity: vp.vector = vp.vector(*velocity, 0)
 
 
 class BallParameters:
-    def __init__(self, physics, color=vp.color.red, name=''):
-        self.mass = physics.mass
-        self.position = physics.position
-        self.velocity = physics.velocity
-        self.color = color
-        self.name = name
+    def __init__(self, physics: PhysicsParameters, color: vp.vector = vp.color.red, name: str = ''):
+        self.mass: float = physics.mass
+        self.position: vp.vector = physics.position
+        self.velocity: vp.vector = physics.velocity
+        self.color: vp.vector = color
+        self.name: str = name
 
 
 class Ball:
-    def __init__(self, params):
-        self.mass = params.mass
-        self.position = params.position
-        self.velocity = params.velocity
+    def __init__(self, params: BallParameters):
+        self.mass: float = params.mass
+        self.position: vp.vector = params.position
+        self.velocity: vp.vector = params.velocity
 
         # Radius proportional to the mass (0.5m for 1 kg)
-        self.radius = 0.5 * (self.mass ** (1/3))
+        self.radius: float = 0.5 * (self.mass ** (1/3))
 
-        self.name = params.name
-        self.sphere = vp.sphere(pos=self.position,
-                                radius=self.radius,
-                                color=params.color,
-                                make_trail=True)
-        self.label = vp.label(pos=self.position,
-                              text=params.name,
-                              height=14,
-                              color=vp.color.white,
-                              box=False,
-                              opacity=0)
+        self.name: str = params.name
+        self.sphere: vp.sphere = vp.sphere(pos=self.position,
+                                           radius=self.radius,
+                                           color=params.color,
+                                           make_trail=True)
+        self.label: vp.label = vp.label(pos=self.position,
+                                        text=params.name,
+                                        height=14,
+                                        color=vp.color.white,
+                                        box=False,
+                                        opacity=0)
 
     @property
-    def angle(self):
+    def angle(self) -> float:
         return vp.degrees(vp.atan2(self.velocity.y, self.velocity.x))
 
     @property
-    def speed(self):
+    def speed(self) -> float:
         return vp.mag(self.velocity)
 
     @property
-    def momentum(self):
+    def momentum(self) -> vp.vector:
         return self.velocity * self.mass
 
     @property
-    def momentum_mag(self):
+    def momentum_mag(self) -> float:
         return vp.mag(self.momentum)
 
-    def update_position(self, dt):
+    def update_position(self, dt: float) -> None:
         self.position += self.velocity * dt
         self.sphere.pos = self.position
         self.label.pos = self.position
@@ -80,44 +80,46 @@ class IntersectionInfo:
 
 
 class BallCollisionSimulator:
-    def __init__(self, ball1_params, ball2_params, simulation_time):
-        self.ball1_params = ball1_params
-        self.ball2_params = ball2_params
-        self.simulation_time = simulation_time
-        self.scene = None
-        self.ball1 = None
-        self.ball2 = None
-        self.ball1_state_t0 = None
-        self.ball2_state_t0 = None
-        self.initial_distance = None
-        self.total_momentum = None
-        self.dot_product = None
-        self.relative_speed = None
-        self.collision_info = None
-        self.intersection_info = None
+    def __init__(self, ball1_params: BallParameters,
+                 ball2_params: BallParameters,
+                 simulation_time: float):
+        self.ball1_params: BallParameters = ball1_params
+        self.ball2_params: BallParameters = ball2_params
+        self.simulation_time: float = simulation_time
+        self.scene: Optional[vp.canvas] = None
+        self.ball1: Optional[Ball] = None
+        self.ball2: Optional[Ball] = None
+        self.ball1_state_t0: Optional[Ball] = None
+        self.ball2_state_t0: Optional[Ball] = None
+        self.initial_distance: Optional[float] = None
+        self.total_momentum: Optional[vp.vector] = None
+        self.dot_product: Optional[float] = None
+        self.relative_speed: Optional[float] = None
+        self.collision_info: Optional[CollisionInfo] = None
+        self.intersection_info: Optional[IntersectionInfo] = None
 
     @staticmethod
-    def create_simulator(ball1_specs, ball2_specs, simulation_time):
-        ball1_physics = PhysicsParameters(*ball1_specs)
-        ball1_params = BallParameters(ball1_physics, color=vp.color.blue, name='1')
-
-        ball2_physics = PhysicsParameters(*ball2_specs)
-        ball2_params = BallParameters(ball2_physics, color=vp.color.red, name='2')
+    def create_simulator(phys1_params: PhysicsParameters,
+                         phys2_params: PhysicsParameters,
+                         simulation_time: float) -> 'BallCollisionSimulator':
+        ball1_params : BallParameters = BallParameters(phys1_params, color=vp.color.blue, name='1')
+        ball2_params : BallParameters = BallParameters(phys2_params, color=vp.color.red, name='2')
 
         return BallCollisionSimulator(ball1_params, ball2_params, simulation_time)
 
-    def remove_scene(self):
-        self.scene.delete()
-        self.scene = None
+    def remove_scene(self) -> None:
+        if self.scene:
+            self.scene.delete()
+            self.scene = None
 
-    def quit_simulation(self):
+    def quit_simulation(self) -> None:
         # We don't import vp_services until needed, because importing it will start
         # the server, if not started already.
         import vpython.no_notebook as vp_services
 
         vp_services.stop_server()
 
-    def __init_simulation(self):
+    def __init_simulation(self) -> None:
         self.scene = vp.canvas(title='Elastic Collision Simulation',
                                width=800,
                                height=800,
@@ -141,15 +143,16 @@ class BallCollisionSimulator:
         # Calculate initial distance
         self.initial_distance = vp.mag(
             self.ball1.position - self.ball2.position)
-        
+
         # Calculate the relative speed of the balls to each other
         self.dot_product = vp.dot(self.ball1.velocity - self.ball2.velocity, 
                                   self.ball1.position - self.ball2.position)
         self.relative_speed = (self.ball1.velocity - self.ball2.velocity).mag
 
-    def __create_grid_and_axes(self):
-        GRID_RANGE = 10
-        step = 1
+    def __create_grid_and_axes(self) -> None:
+        GRID_RANGE : int = 10
+        step : int = 1
+
         for x in vp.arange(-GRID_RANGE, GRID_RANGE + step, step):
             vp.curve(pos=[vp.vector(x, -GRID_RANGE, 0),
                           vp.vector(x, GRID_RANGE, 0)],
@@ -166,7 +169,7 @@ class BallCollisionSimulator:
                  text='Y', height=16, box=False)
 
     @staticmethod
-    def __print_velocity_details(ball1, ball2):
+    def __print_velocity_details(ball1: Ball, ball2: Ball) -> None:
         print()
         for ball in ball1, ball2:
             print(f'Ball {ball.name}:')
@@ -180,7 +183,10 @@ class BallCollisionSimulator:
                 ball.momentum_mag:.2f} N-s at {ball.angle:.2f} degrees')
         print()
 
-    def __calculate_intersection(self):
+    def __calculate_intersection(self) -> None:
+        assert self.ball1_state_t0
+        assert self.ball2_state_t0
+
         # Extract initial positions and velocities
         x1, y1 = self.ball1_state_t0.position.x, self.ball1_state_t0.position.y
         x2, y2 = self.ball2_state_t0.position.x, self.ball2_state_t0.position.y
@@ -201,9 +207,9 @@ class BallCollisionSimulator:
             return None  # Lines are parallel
 
         t = ((x1 - x2) * (y2 - y2_end) - (y1 - y2)
-             * (x2 - x2_end)) / denominator
+            * (x2 - x2_end)) / denominator
         u = -((x1 - x1_end) * (y1 - y2) -
-              (y1 - y1_end) * (x1 - x2)) / denominator
+            (y1 - y1_end) * (x1 - x2)) / denominator
 
         # Correct -0.0 to 0.0
         if t == -0.0:
@@ -218,33 +224,43 @@ class BallCollisionSimulator:
 
             # Store intersection info
             self.intersection_info = IntersectionInfo(position=vp.vector(ix, iy, 0),
-                                                      ball1_time=t * self.simulation_time,
-                                                      ball2_time=u * self.simulation_time)
+                                                    ball1_time=t * self.simulation_time,
+                                                    ball2_time=u * self.simulation_time)
 
-    def __process_post_collision_physics(self):
-        m1 = self.ball1.mass
-        m2 = self.ball2.mass
-        v1_initial = self.ball1.velocity
-        v2_initial = self.ball2.velocity
+    def __process_post_collision_physics(self) -> None:
+        assert self.ball1
+        assert self.ball2
+
+        m1 : float = self.ball1.mass
+        m2 : float = self.ball2.mass
+        v1_initial : vp.vector = self.ball1.velocity
+        v2_initial : vp.vector = self.ball2.velocity
         self.ball1.velocity = (v1_initial * (m1 - m2) +
-                               2 * m2 * v2_initial) / (m1 + m2)
+                            2 * m2 * v2_initial) / (m1 + m2)
         self.ball2.velocity = (v2_initial * (m2 - m1) +
-                               2 * m1 * v1_initial) / (m1 + m2)
+                            2 * m1 * v1_initial) / (m1 + m2)
 
-    def __verify_conservation_of_momentum(self):
+    def __verify_conservation_of_momentum(self) -> None:
+        assert self.ball1
+        assert self.ball2
+        assert self.total_momentum
+
         # Calculate final total momentum
-        final_total_momentum = self.ball1.momentum + self.ball2.momentum
+        final_total_momentum : vp.vector = self.ball1.momentum + self.ball2.momentum
 
         # Verify momentum has been conserved
         assert round(self.total_momentum.mag, ndigits=3) == \
-               round(final_total_momentum.mag, ndigits=3), \
-               f'Initial total: {self.total_momentum.mag}, Final total: {final_total_momentum.mag}'
-    
-    def __run_simulation(self):
-        dt = 0.01
-        time_elapsed = 0
+            round(final_total_momentum.mag, ndigits=3), \
+            f'Initial total: {self.total_momentum.mag}, Final total: {final_total_momentum.mag}'
 
-        SIMULATION_TIME_AFTER_COLLISION = 2  # secs
+    def __run_simulation(self) -> None:
+        assert self.ball1
+        assert self.ball2
+
+        dt : float = 0.01
+        time_elapsed : float = 0.0
+
+        SIMULATION_TIME_AFTER_COLLISION : int = 2  # secs
 
         while True:
             vp.rate(100)
@@ -253,7 +269,7 @@ class BallCollisionSimulator:
             # distance of both radiuses, we have a collision
             if not self.collision_info and \
                     vp.mag(self.ball1.position - self.ball2.position) <= \
-                          (self.ball1.radius + self.ball2.radius):
+                        (self.ball1.radius + self.ball2.radius):
                 # update balls based on physics of collision
                 self.__process_post_collision_physics()
 
@@ -273,8 +289,15 @@ class BallCollisionSimulator:
             self.ball2.update_position(dt)
             time_elapsed += dt
 
-    def run(self):
+    def run(self) -> None:
         self.__init_simulation()
+
+        assert self.ball1
+        assert self.ball2
+        assert self.ball1_state_t0
+        assert self.ball2_state_t0
+        assert self.dot_product
+        assert self.total_momentum
 
         print('***************************************************')
         print('Initial Conditions:')
@@ -309,11 +332,11 @@ class BallCollisionSimulator:
             if self.intersection_info:
                 print('Paths did intersect though:')
                 print(f'  Path Intersection Point: ({self.intersection_info.position.x:.2f}, {
-                      self.intersection_info.position.y:.2f})')
+                    self.intersection_info.position.y:.2f})')
                 print(f'  Time for Ball 1 to reach intersection: {
-                      self.intersection_info.ball1_time:.2f} secs')
+                    self.intersection_info.ball1_time:.2f} secs')
                 print(f'  Time for Ball 2 to reach intersection: {
-                      self.intersection_info.ball2_time:.2f} secs')
+                    self.intersection_info.ball2_time:.2f} secs')
             else:
                 print('No path intersection found either.')
 
@@ -322,54 +345,56 @@ class BallCollisionSimulator:
         self.__verify_conservation_of_momentum()
 
 
-def get_user_input():
-    def get_float(prompt):
-        while True:
-            try:
-                return float(input(prompt))
-            except ValueError:
-                print("Please enter a valid number.")
-
-    def get_vector(prompt):
-        while True:
-            try:
-                x, y = map(float, input(prompt).split(','))
-                return (x, y)
-            except ValueError:
-                print("Please enter two numbers separated by a comma.")
-
-    print("Enter parameters for Ball 1:")
-    mass1 = get_float("Mass (kg): ")
-    position1 = get_vector("Position (x,y) in meters: ")
-    velocity1 = get_vector("Velocity (x,y) in m/s: ")
-
-    print("\nEnter parameters for Ball 2:")
-    mass2 = get_float("Mass (kg): ")
-    position2 = get_vector("Position (x,y) in meters: ")
-    velocity2 = get_vector("Velocity (x,y) in m/s: ")
-
-    simulation_time = get_float("\nEnter simulation time (seconds): ")
-
-    return (mass1, position1, velocity1), (mass2, position2, velocity2), simulation_time
-
-
 if __name__ == '__main__':
+    def get_user_input() -> Tuple[PhysicsParameters, PhysicsParameters, float]:
+        def get_float(prompt: str) -> float:
+            while True:
+                try:
+                    return float(input(prompt))
+                except ValueError:
+                    print("Please enter a valid number.")
+
+        def get_vector(prompt: str) -> Tuple[float, float]:
+            while True:
+                try:
+                    x, y = map(float, input(prompt).split(','))
+                    return (x, y)
+                except ValueError:
+                    print("Please enter two numbers separated by a comma.")
+
+        print("Enter parameters for Ball 1:")
+        mass1 : float = get_float("Mass (kg): ")
+        position1 : vp.vector = get_vector("Position (x,y) in meters: ")
+        velocity1 : vp.vector = get_vector("Velocity (x,y) in m/s: ")
+
+        print("\nEnter parameters for Ball 2:")
+        mass2 : float = get_float("Mass (kg): ")
+        position2 : vp.vector = get_vector("Position (x,y) in meters: ")
+        velocity2 : vp.vector = get_vector("Velocity (x,y) in m/s: ")
+
+        simulation_time : float = get_float("\nEnter simulation time (seconds): ")
+
+        return (PhysicsParameters(mass1, position1, velocity1),
+                PhysicsParameters(mass2, position2, velocity2),
+                simulation_time)
+
     parser = argparse.ArgumentParser(description='Ball Collision Simulator')
     parser.add_argument('--test', action='store_true', help='Run with pre-defined test case')
     args = parser.parse_args()
 
     if args.test:
         # Pre-defined test case
-        ball1_specs = (1.0, (3.0, 4.0), (-1.0, -0.5))
-        ball2_specs = (1.0, (-3.3, -4.0), (0.0, 1.0))
-        simulation_time = 10.0
+        #                                Mass, ( Position )  ( Velocity )
+        ball1_params = PhysicsParameters( 1.0, ( 3.0,  4.0), (-3.0, -0.1))
+        ball2_params = PhysicsParameters( 4.0, (-3.3, -4.0), ( 0.0,  4.0))
+        simulation_time: float = 10.0  # secs
     else:
         # Get user input
-        ball1_specs, ball2_specs, simulation_time = get_user_input()
+        ball1_params, ball2_params, simulation_time = get_user_input()
 
-    ball_collision_sim = BallCollisionSimulator.create_simulator(
-        ball1_specs,
-        ball2_specs,
+    ball_collision_sim: BallCollisionSimulator = BallCollisionSimulator.create_simulator(
+        ball1_params,
+        ball2_params,
         simulation_time
     )
